pyproject.toml
[project]
name = "FoundationPose"
author = "Joshua Powers"
version = "0.1.0"
description = "A remake of the foundation pose pipeline and tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cupy-cuda12x>=13.5.1",
    "diffrp-nvdiffrast>=0.3.3.1",
    "imageio>=2.37.0",
    "kornia>=0.8.1",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "scipy>=1.16.1",
    "tensorrt>=10.13.0.35",
    "torch>=2.7.1",
    "trimesh>=4.7.1",
]
